change RandomGenerator class's name
